Second attempt to bring track gain and panning as undoable commands, now succeeded.
git-svn-id: http://svn.code.sf.net/p/qtractor/code_attic/trunk@305 304e7ffa-9037-4501-99c0-529757c9bbcc

+++ src/qtractorMixer.h
@@ -101,8 +101,8 @@ protected slots:
 	void busButtonSlot();
 
 	// Meter slider change slots.
-	void panChangedSlot(int);
-	void gainChangedSlot(int);
+	void panChangedSlot(float);
+	void gainChangedSlot(float);
 
 protected:
